+++ src/partition/VerticalPartition.java
@@ -1,19 +1,17 @@
 package partition;
 
-import lombok.AllArgsConstructor;
-
 import java.util.Random;
 
 public class VerticalPartition implements PartitionFunction {
-    private final float xBoundary;
+    private final double xBoundary;
 
-    public VerticalPartition(float scaleSize) {
+    public VerticalPartition(double scaleSize) {
         Random random = new Random();
-        xBoundary = (float)random.nextGaussian(0, scaleSize/2);
+        xBoundary = random.nextGaussian(0, scaleSize/2);
         System.out.println("boundary at x = " + xBoundary);
     }
     @Override
-    public int getPartition(float x, float y) {
+    public int getPartition(double x, double y) {
         if (x < xBoundary) {
             return 0;
         }
